// src/app/components/tms/model/common.model.ts
export class EmployeeModel {
    employeeId: string;
    employeeName: string;

}

export class DepartmentModel {
    departmentId: string;
    departmentName: string;

}

export class DivisionModel {
    divisionId: string;
    divisionName: string;

}

export class ReferenceModel {

    referenceId: number;
    referenceType: string;
    referenceSourceId: string;
    referenceNumber: string;
    referenceNumberName: string;
}

export class DashboardModel {

    assignmentId: number;
    registrationNumber: string;
    driverName: string;
    usedType: string;
    pickupLocation: string;
    dropLocation: string;
    assignStartTime: string;
    assignEndTime: string;
    assignStartHour: string;
    assignEndHour: string;
    assignHours: string;
    oneAm: string;


    oneThirtyAm: string;
    twoAm: string;
    twoThirtyAm: string;
    threeAm: string;
    threeThirtyAm: string;
    fourAm: string;
    fourThirtyAm: string;
    fiveAm: string;
    fiveThirtyAm: string;
    sixAm: string;
    sixThirtyAm: string;
    sevenAm: string;
    sevenThirtyAm: string;
    eightAm: string;
    eightThirtyAm: string;
    nineAm: string;




    nineThirtyAm: string;
    tenAm: string;
    tenThirtyAm: string;
    elevenAm: string;
    elevenThirtyAm: string;
    twelveAm: string;
    twelveThirtyAm: string;
    onePm: string;
    oneThirtyPm: string;
    twoPm: string;
    twoThirtyPm: string;
    threePm: string;
    threeThirtyPm: string;
    fourPm: string;
    fourThirtyPm: string;

    fivePm: string;
    fiveThirtyPm: string;
    sixPm: string;
    sixThirtyPm: string;
    sevenPm: string;
    sevenThirtyPm: string;
    eightPm: string;
    eightThirtyPm: string;
    ninePm: string;
    nineThirtyPm: string;
    tenPm: string;
    tenThirtyPm: string;
    elevenPm: string;
    elevenThirtyPm: string;
}